Added *.log to gitignore Updated Readme with Markdown as well as further detailed setup instructions to further match the changes
Redesigned the core system to utilize a method of Hooks to load core files. This was to separate the code from the main file for less clutter allowing code to be separated into areas.
Redesigned config file and variables to match and function for the new system better. All variables are reduced to the base $cfg for easier inclusion for globals rather than needing multiple.
Improved Authentication System the system nows for specific access authentication that can be configured per command or per command/subcommands. Access can be given based upon GroupID or UniqueID. Unauthorized Access is logged to bot_security.log regardless of if logging is enabled.
Improved Message/Debug/Log handling it is now possible to log the bot, or enable debug messages without parsing the debug to the console. Debug logging now outputs to bot_debug.log, as well as enabling debugging will automatically turn on PHP error_logging which logs to php_error.log.
The default $cfg['log'] need not be on as it all other normal messages. Security, Debug, and PHP error logs will log without this on.
Fixed bug found with command_auth that caused access to fail if you only had one ServerGroup.